refactor share query to its own component
