+++ src/app/views/query-runner/query-input/QueryInput.tsx
@@ -1,24 +1,19 @@
-import { IDropdownOption, Dropdown, Dialog, IconButton, DialogType,
-  FontSizes, DialogFooter, DefaultButton, IIconProps, TooltipHost, DirectionalHint } from '@fluentui/react';
-import React, { useEffect, useState } from 'react';
+import { IDropdownOption, Dropdown } from '@fluentui/react';
+import React from 'react';
 import { injectIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
-import { componentNames, eventTypes, telemetry } from '../../../../telemetry';
 import { httpMethods, IQueryInputProps } from '../../../../types/query-runner';
 
 import { IRootState } from '../../../../types/root';
 import { setSampleQuery } from '../../../services/actions/query-input-action-creators';
 import { GRAPH_API_VERSIONS } from '../../../services/graph-constants';
 import { getStyleFor } from '../../../utils/http-methods.utils';
-import { sanitizeQueryUrl } from '../../../utils/query-url-sanitization';
 import { parseSampleUrl } from '../../../utils/sample-url-generation';
 import { translateMessage } from '../../../utils/translate-messages';
 import SubmitButton from '../../../views/common/submit-button/SubmitButton';
-import { copy } from '../../common/copy';
-import { CopyButton } from '../../common/copy/CopyButton';
-import { createShareLink } from '../../common/share';
 import { queryRunnerStyles } from '../QueryRunner.styles';
 import { AutoComplete } from './auto-complete';
+import { ShareQuery } from './share-query';
 
 
 const QueryInput = (props: IQueryInputProps) => {
@@ -44,35 +39,12 @@ const QueryInput = (props: IQueryInputProps) => {
 
   const { mobileScreen } = sidebarProperties;
 
-  const [showShareQueryDialog, setShareQuaryDialogStatus] = useState(true);
-  const [shareLink, setShareLink] = useState(() => createShareLink(sampleQuery));
-
-  const toggleShareQueryDialogState = () => {
-    setShareQuaryDialogStatus(prevState => !prevState);
-  };
-
-  useEffect(() => {
-    setShareLink(createShareLink(sampleQuery));
-  }, [sampleQuery]);
-
-  const iconProps : IIconProps = {
-    iconName: 'Share'
-  }
-
   const showError = !authenticated && sampleQuery.selectedVerb !== 'GET';
   const verbSelector: any = queryRunnerStyles().verbSelector;
   verbSelector.title = {
     ...verbSelector.title,
     background: getStyleFor(sampleQuery.selectedVerb)
   };
-
-  const shareButtonStyles = queryRunnerStyles().iconButton;
-
-  const calloutProps = {
-    gapSpace: 0
-  };
-
-  const content = <div style={{padding:'3px'}}>{translateMessage('Share Query')}</div>
 
   const contentChanged = (value: string) => {
     const query = { ...sampleQuery, ...{ sampleUrl: value } };
@@ -104,20 +76,6 @@ const QueryInput = (props: IQueryInputProps) => {
       handleOnRunQuery();
     }, 500);
   };
-
-  const handleCopy = () => {
-    copy('share-query-text');
-    trackCopyEvent();
-  };
-
-  const trackCopyEvent = () => {
-    const sanitizedUrl = sanitizeQueryUrl(sampleQuery.sampleUrl);
-    telemetry.trackEvent(eventTypes.BUTTON_CLICK_EVENT,
-      {
-        ComponentName: componentNames.SHARE_QUERY_COPY_BUTTON,
-        QuerySignature: `${sampleQuery.selectedVerb} ${sanitizedUrl}`
-      });
-  }
 
   return (
     <>
@@ -160,17 +118,7 @@ const QueryInput = (props: IQueryInputProps) => {
             />
           </div>
           <div className='col-lg-1' style={{flexShrink: 2}}>
-            <TooltipHost
-              content={content}
-              calloutProps={calloutProps}
-              directionalHint={DirectionalHint.leftBottomEdge}
-            >
-              <IconButton
-                onClick={toggleShareQueryDialogState}
-                iconProps={iconProps}
-                styles={shareButtonStyles}
-              />
-            </TooltipHost>
+            <ShareQuery/>
           </div>
         </>
         }
@@ -189,54 +137,10 @@ const QueryInput = (props: IQueryInputProps) => {
           />
         </div>
         <div style={{flexGrow: 1, flexShrink: 2}}>
-          <TooltipHost
-            content={content}
-            calloutProps={calloutProps}
-            directionalHint={DirectionalHint.leftBottomEdge}
-          >
-            <IconButton
-              onClick={toggleShareQueryDialogState}
-              iconProps={iconProps}
-              styles={shareButtonStyles}
-            />
-          </TooltipHost>
+          <ShareQuery/>
         </div>
       </div>
       }
-      <Dialog
-        hidden={showShareQueryDialog}
-        onDismiss={toggleShareQueryDialogState}
-        dialogContentProps={{
-          type: DialogType.normal,
-          title: 'Share Query',
-          isMultiline: true,
-          subText: translateMessage('Share Query Message')
-        }}
-      >
-        <textarea
-          style={{
-            wordWrap: 'break-word',
-            fontFamily: 'monospace',
-            fontSize: FontSizes.xSmall,
-            width: '100%',
-            height: 63,
-            overflowY: 'scroll',
-            resize: 'none',
-            color: 'black'
-          }}
-          id='share-query-text'
-          className='share-query-params'
-          defaultValue={shareLink}
-          aria-label={translateMessage('Share Query')}
-        />
-        <DialogFooter>
-          <CopyButton handleOnClick={handleCopy} isIconButton={false} />
-          <DefaultButton
-            text={translateMessage('Close')}
-            onClick={toggleShareQueryDialogState}
-          />
-        </DialogFooter>
-      </Dialog>
     </>
   )
 }
