+++ src/app/views/query-runner/query-input/share-query/ShareQuery.styles.ts
@@ -1,0 +1,11 @@
+export const shareQueryStyles = () => {
+  return {
+    iconButton: {
+      root: {
+        float: 'right',
+        border: '1px solid',
+        width: '100%'
+      }
+    }
+  }
+}
