+++ src/app/views/query-runner/query-input/share-query/index.ts
@@ -1,0 +1,1 @@
+export { ShareQuery } from './ShareQuery';
